backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ezkl = "^2.4.4"
flask-cors = "^4.0.0"
torch = "^2.1.0"
torchvision = "^0.16.0"
pillow = "^10.0.1"
flask = {extras = ["async"], version = "^3.0.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
